requirements.txt
aiokafka
cassandra-driver
PyYAML
pandas
